Move Parameters and QualifiedParameters in their own files
